+++ yandex/resource_yandex_container_repository_iam_binding.go
@@ -1,0 +1,7 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+
+func resourceYandexContainerRepositoryIAMBinding() *schema.Resource {
+	return resourceIamBindingWithImport(IamContainerRepositorySchema, newContainerRepositoryIamUpdater, containerRepositoryIDParseFunc)
+}